fixed charge attack
the charge attack will now be initiated when the mouse is a set distance
from the edge of the screen. it also has two timers, one for how long it
lasts and one for a cooldown between uses. also created some basic UI to
display and control the speed of the player character while playing.