dash fix and civilian health system
fixed the hero's dash so that it is cancelled if the mouse leaves the
threshold for it. set up the beginnings of a health system for
civilians: they start with 100 health, which will decrease if they get
damaged (for now it happens if you press D). If their health reaches 0
they go into a "dying" state, which causes them to stop moving and
receiving damage for 30 seconds. After that time is up, they are "dead"
and are destroyed.