begin shift toward rts style unit movement
created a new script to handle the setting of destinations for the civs
currently selected. each civ needs a separate destination in order to
solve the current movement problems. this script will keep track of
which civs are selected, the placement of the destination and marker,
and where each specific civ should go. a new tag called "Ground" was
added in order for the new marker placement to work. now that the
civilian controller no longer needs to handle destination creation, it
has been removed (currently temporarily by commenting out). the
controller does, however, need to let the new script know which civ is
selected, so that capability has been added. NOTE: only the framework of
the new script has been created; it does nothing more than place a
marker right now. so civs can't actually move in this version.