new civ movement and some fixes
fully implemented the new civilian movement system. it requires that the
CivRTSUnitHandling script be placed on something in the scene (i.e. the
camera) and then dragged onto each civilian's script in the inspector.
to place a marker, at least one civ has to be selected and "Ground" has
to be directly below the mouse, with nothing else in-between. once a
marker is placed, the selected civ(s) will go toward it, with the
closest civ actually reaching that point and all other civs going to a
point relative to the first civ. the issue of them heading back to their
previous position after being forced to move has been resolved
differently, with the use of the boolean 'stop'. the stop control code
has been changed, and extraneous code deleted. the way the whole system
works is outlined by comments in the CivRTSUnitHandling script.