menu template + more civ fixes
downloaded the game jam menu template from the asset store and will be
using it for the menus in the game. made civs stop when they get damaged
(in preparation for a knockback animation) and then start moving again
once their invincibility runs out. fixed a bug where, when dead, a
selected civ would be destroyed but still be referenced in the
selectedCivs list, thus throwing an error. also commented out
unnecessary debug logs.