pickup + civilian count + odds and ends
created the 'first aid' pickup, which allows the player to fully revive
a 'Dying' civ. first the player needs to pick it up by going near it,
then they need to get close to a civ in the 'Dying' state. player can
only have one pickup at a time. added a script to manage overall stuff,
that wouldn't belong in any other script. right now it keeps track of
the amount of 'activated' and 'alive' civs in the level, as well as the
pickup. the script is on an object under the player object, with a
trigger for detecting civs/pickups. also did a bit more tweaking of the
civilian and player controller scripts - including allowing play in the
editor without going through the menu first, and making sure only
activated civs can be controlled.