hero facing direction fix
did not even notice that this feature no longer worked; it hasn't worked
since the switch to mouse controls. fixed it so now the hero will change
which way they're facing based on where the mouse is. also made civvies
invincible when they are brought back from the dying state.