civilian activation change
changed it so that civilians become active once they are seen by the
camera, or rather, any camera. this means scene view cameras as well, so
if you want to test it properly make sure to move the scene view to an
empty part of the level. also deleted the old pickup move code.