level end and moving platforms
made a trigger for the end of the level, that will count up the gathered
civs and (eventually) present a ui window to show the results to the
player. it also disables control of civs once they pass through it. the
moving platform now uses itween for node placement but script for actual
movement. for circle movement it uses a separate circle object
controlled by script alone. the moving platform can only have one
movement type at a time.